Update: Delete category

+++ src/categories/categories.service.ts
@@ -1,28 +1,30 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Inject, Injectable, forwardRef } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CategoryDTO } from './dto/category.dto';
 import { Category } from './entities/category.entity';
+import { ProductsService } from 'src/products/products.service';
 @Injectable()
 export class CategoriesService {
   constructor(
     @InjectRepository(Category) private categoryRepo: Repository<Category>,
+    @Inject(forwardRef(() => ProductsService))
+    private productService: ProductsService,
   ) {}
   create(createCategoryDto: CategoryDTO) {
-
-   return  this.categoryRepo.save(createCategoryDto)
+    return this.categoryRepo.save(createCategoryDto);
   }
 
   findAll() {
     return this.categoryRepo.find();
   }
   async findOne(id: number) {
-     
+   
     const category = await this.categoryRepo.findBy({
-      id : id
+      id: id,
     });
 
-    return category
+    return category;
   }
 
   update(id: number, updateCategoryDto: Category) {
@@ -30,14 +32,19 @@ export class CategoriesService {
   }
 
   async remove(id: number) {
-    const category = await this.findOne(id)
-    if(category.length){
-      return this.categoryRepo.delete({
-        id : id
-      })
+    const category = await this.findOne(id);
+     const product = this.productService.findByCategory(id);
+     if ((await product).length) {
+       return new HttpException('Product  found', HttpStatus.FOUND);
+     }else{
+       if (category.length) {
+         return this.categoryRepo.delete({
+           id: id,
+         });
+       } else {
+         return new HttpException('Category not found', HttpStatus.NOT_FOUND);
+       }
 
-    }else{
-       return new HttpException('Category not found', HttpStatus.NOT_FOUND);
-    }
+     }
   }
 }
